chore: move all deprecated fields that are only used by gjafakort and delete non-used one
