+++ apps/gjafakort/web/components/FieldInput/FieldInput.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
 import { FormikState, FieldInputProps } from 'formik'
 import get from 'lodash/get'
-import { Input } from '../Input/Input'
-import { useDeprecatedComponent } from '../private/useDeprecatedComponent'
+import { Input } from '@island.is/island-ui/core'
 
 interface FieldNumberInputProps {
   form: FormikState<string | number>
@@ -14,7 +13,6 @@ export const FieldInput = ({
   form: { touched, errors },
   ...props
 }: FieldNumberInputProps) => {
-  useDeprecatedComponent('FieldInput')
   const nameArray = (field.name && field.name.split('.')) || []
   return (
     <Input
