+++ apps/gjafakort/web/components/FieldSelect/FieldSelect.tsx
@@ -2,8 +2,7 @@ import React from 'react'
 import { FieldInputProps, FormikState } from 'formik'
 import { ValueType } from 'react-select'
 import get from 'lodash/get'
-import { Select, SelectProps, Option } from '../Select/Select'
-import { useDeprecatedComponent } from '../private/useDeprecatedComponent'
+import { Select, SelectProps, Option } from '@island.is/island-ui/core'
 
 export interface FieldSelectProps extends SelectProps {
   form: FormikState<string | number>
@@ -15,7 +14,6 @@ export const FieldSelect = ({
   form: { touched, errors },
   ...props
 }: FieldSelectProps) => {
-  useDeprecatedComponent('FieldSelect')
   const nameArray = (field.name && field.name.split('.')) || []
   return (
     <Select
