+++ apps/gjafakort/web/screens/Companies/Application/components/SelectionForm/SelectionForm.tsx
@@ -2,7 +2,6 @@ import React from 'react'
 import { Formik, Form, Field } from 'formik'
 
 import {
-  FieldSelect,
   Box,
   Column,
   Columns,
@@ -13,6 +12,7 @@ import {
 
 import { useI18n } from '@island.is/gjafakort-web/i18n'
 import { Company } from '@island.is/gjafakort-web/graphql/schema'
+import { FieldSelect } from '@island.is/gjafakort-web/components'
 
 interface PropTypes {
   onSubmit: (values: Company) => void
