+++ apps/gjafakort/web/screens/User/components/Barcode/Barcode.tsx
@@ -16,12 +16,14 @@ import {
   Column,
   SkeletonLoader as SL,
   Divider,
-  FieldNumberInput,
-  FieldInput,
 } from '@island.is/island-ui/core'
 
 import { useI18n } from '@island.is/gjafakort-web/i18n'
-import { ErrorPanel } from '@island.is/gjafakort-web/components'
+import {
+  ErrorPanel,
+  FieldInput,
+  FieldNumberInput,
+} from '@island.is/gjafakort-web/components'
 
 import { barcodeMachine } from './barcodeMachine'
 import { Countdown, RenderBarcode } from '..'
