+++ libs/island-ui/core/src/index.ts
@@ -10,7 +10,6 @@ export * from './lib/Page/Page' // TODO: missing stories
 export * from './lib/BulletList/BulletList'
 export * from './lib/VideoIframe/VideoIframe' // TODO: missing stories
 export * from './lib/Tooltip/Tooltip'
-export * from './lib/FieldPolarQuestion/FieldPolarQuestion' // TODO: missing stories
 export * from './lib/LinkCard/LinkCard'
 export * from './lib/Tag/Tag'
 export * from './lib/SkeletonLoader/SkeletonLoader'
@@ -34,7 +33,7 @@ export * from './lib/Icon/Icon' // TODO: missing stories
 export * from './lib/Typography/Typography'
 export * from './lib/Text/Text'
 
-// Forms
+// Form
 export * from './lib/AsyncSelect/AsyncSelect'
 export * from './lib/Checkbox/Checkbox'
 export * from './lib/RadioButton/RadioButton'
@@ -43,11 +42,6 @@ export * from './lib/Select/Select'
 export * from './lib/Input/Input'
 export * from './lib/InputFileUpload/InputFileUpload'
 export * from './lib/InputError/InputError' // TODO: missing stories
-export * from './lib/FieldCheckbox/FieldCheckbox' // TODO: missing stories
-export * from './lib/FieldNumberInput/FieldNumberInput' // TODO: missing stories
-export * from './lib/FieldInput/FieldInput' // TODO: missing stories
-export * from './lib/FieldSelect/FieldSelect' // TODO: missing stories
-export * from './lib/FieldRadioButton/FieldRadioButton' // TODO: missing stories
 
 // Layout
 export * from './lib/Box/Box'
